% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classes}
\alias{classes}
\title{Data set from Economist article}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 40 rows and 9 columns.}
\usage{
classes
}
\description{
Data set from Economist article
}
\keyword{datasets}
