% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{exdata}
\alias{exdata}
\alias{roxy_tag_parse.roxy_tag_exdata}
\alias{roxy_tag_rd.roxy_tag_exdata}
\title{Document Dataset Generation}
\usage{
\method{roxy_tag_parse}{roxy_tag_exdata}(x)

\method{roxy_tag_rd}{roxy_tag_exdata}(x, base_path, env)
}
\arguments{
\item{x}{The tag}

\item{base_path}{Path to package root directory.}

\item{env}{Environment in which to evaluate code (if needed)}
}
\value{
\code{roxy_tag_parse}: A \code{\link[roxygen2]{roxy_tag}} object with
a value being a two-length list:
\describe{
 \item{\code{name}}{Name of the dataset}
 \item{\code{text}}{Code to generate the dataset}
}

\code{roxy_tag_rd}: An \code{\link[roxygen2]{rd_section}} object of
type \dQuote{examples} with the value being the code to generate the dataset
}
\description{
Document Dataset Generation
}
\examples{
#' Dataset Documentation File
#'
#' Simply add an `@exdata` tag to the Roxygen documentation for a dataset.
#' For its one and only argument, it takes the name of the data-raw file
#' used to generate the dataset, without extensions
#'
#' @format A dataset
#'
#' @exdata dataset
#'
"dataset"

}
