% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{link}
\alias{link}
\title{Link data for Corona virus app}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 143 rows and 5 columns.}
\usage{
link
}
\description{
Link data for Corona virus app
}
\keyword{datasets}
