% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nycParkingSmall}
\alias{nycParkingSmall}
\title{Mini NYC parking data set}
\format{An object of class \code{data.frame} with 30000 rows and 45 columns.}
\usage{
nycParkingSmall
}
\description{
First 30,000 rows of the parking data.
}
\keyword{datasets}
